+++ app/helpers/hash_helper.py
@@ -4,12 +4,7 @@ from app.config import get_config
 cfg = get_config()
 
 
-class HashHelper:
-    """Hash mixin."""
-
-    @staticmethod
-    def hash(value: str) -> str:
-        """Return hashed value."""
-        encoded_value = (value + cfg.HASH_SALT).encode()
-        hash = hashlib.sha512(encoded_value)
-        return hash.hexdigest()
+def get_hash(value: str) -> str:
+    encoded_value = (value + cfg.HASH_SALT).encode()
+    hash = hashlib.sha512(encoded_value)
+    return hash.hexdigest()